wrap canvas (again!)
